+++ scrapers/toonily/toonily.go
@@ -52,8 +52,8 @@ var Config = &scraper.Configuration{
 	VolumeExtractor: &scraper.VolumeExtractor{
 		// selector that points to only 1 element
 		Selector: "body",
-		Number: func(selection *goquery.Selection) int {
-			return 1
+		Number: func(selection *goquery.Selection) float32 {
+			return 1.0
 		},
 	},
 	ChapterExtractor: &scraper.ChapterExtractor{
